+++ src/Wizard.js
@@ -99,6 +99,10 @@ export default class Wizard extends Webform {
     const onReady = super.init();
     this.setComponentSchema();
 
+    if (this.pages && this.pages.length) {
+      this.component = this.pages[this.page].component;
+    }
+
     this.on('subWizardsUpdated', () => {
       if (this.subWizards.length) {
         this.transformPages();
@@ -613,6 +617,7 @@ export default class Wizard extends Webform {
   }
 
   setValue(submission, flags = {}) {
+    this._submission = submission;
     this.establishPages(submission.data);
     const changed = this.getPages({ all: true }).reduce((changed, page) => {
       return this.setNestedValue(page, submission.data, flags, changed) || changed;