+++ src/components/textarea/TextArea.js
@@ -552,6 +552,7 @@ export default class TextAreaComponent extends TextFieldComponent {
     this.editorsReady = [];
     this.updateSizes.forEach(updateSize => this.removeEventListener(window, 'resize', updateSize));
     this.updateSizes = [];
+    super.detach();
   }
 
   getValue() {